refactor: moved select menus to new system
